Added SystemPreloadScene to preload SceneManager, added StartScene to config.ts
